Simplified participant list pagination by just using the already build in splitMessage function
