+++ src/button/ShowAllParticipantsButton.ts
@@ -1,10 +1,11 @@
 import { ActionRowBuilder, ButtonBuilder, ButtonInteraction, ButtonStyle, MessageFlags } from "discord.js";
-import { assertMessageIsValidParticipantListMessage } from "../permission/assertMessageIsValidParticipantListMessage";
-import { AbstractButton } from "./AbstractButton";
 import { MeetupRow } from "../database/table/Meetup";
-import { getParticipantData } from "../util/meetup/createMeetupInfoEmbed";
-import { createParticipantListPages } from "../util/meetup/createParticipantListPages";
 import { tButton } from "../i18n";
+import { assertMessageIsValidParticipantListMessage } from "../permission/assertMessageIsValidParticipantListMessage";
+import { getParticipantData } from "../util/meetup/createMeetupInfoEmbed";
+import { createParticipantListMessage } from "../util/meetup/createParticipantListMessage";
+import { splitMessage } from "../util/splitMessage";
+import { AbstractButton } from "./AbstractButton";
 
 /**
  * Class for handling "Show all participants" buttonpress in meetup participant list messages
@@ -34,11 +35,12 @@ export class ShowAllParticipantsButton extends AbstractButton {
         const participantData = await getParticipantData(meetup.meetupID);
 
         //get pages
-        const participantPages = createParticipantListPages(participantData);
+        const participantListMessage: string = createParticipantListMessage(participantData);
+        const participantListPages: string[] = splitMessage(participantListMessage);
 
         //components
         const components = [];
-        if(participantPages.length > 1){
+        if(participantListPages.length > 1){
             //add navigation buttons
             const previousPageButton: ButtonBuilder = new ButtonBuilder()
                 .setCustomId("show_participants_switch_page:{" + meetup.meetupID +",0}")
@@ -63,7 +65,7 @@ export class ShowAllParticipantsButton extends AbstractButton {
         }
 
         await interaction.editReply({
-            content: participantPages[0],
+            content: participantListPages[0],
             components: components,
         });
     }
